teste(126184): Corrigindo os testes para o método  update_periodos_alimentacoes
